Emit messages to perspective if required.
If the player that sent the message is in a perspective, send that
message to the perspective, check it against the outgoing filter and
send it to the base room if it passes.

If the player is not in a perspective, check the messages against all
perspective's incoming filter and emit it there if it passes.